From Wang Rui, "The new osgShadow and osgFX serializers are attached, and some modifications of the osgShadow header naming styles as well. The osgDB::Serializer header is also changed to add new Vec2 serializer macros because of the needs of osgShadow classes. It should compile fine on both Windows and Linux. But I have only done a few tests to generate .osgb, .osgt and .osgx formats with these new wrappers."

+++ src/osgPlugins/3ds/lib3ds/lib3ds_file.c
@@ -446,13 +446,13 @@ mdata_read(Lib3dsFile *file, Lib3dsIo *io) {
 
 static int 
 compare_node_id( const void *a, const void *b ) {
-   return (*((Lib3dsNode**)a))->node_id - (*((Lib3dsNode**)b))->node_id;
+   return (int)((*((Lib3dsNode**)a))->node_id) - (int)((*((Lib3dsNode**)b))->node_id);
 }
 
 
 static int 
 compare_node_id2( const void *a, const void *b ) {
-   return *((unsigned short*)a) - (*((Lib3dsNode**)b))->node_id;
+   return (int)(*((unsigned short*)a)) - (int)((*((Lib3dsNode**)b))->node_id);
 }
